CRM Changes 03/12
1.        When in the CRM Opportunity. If you firstly filter for a
particular Sales Rep (use Dean Barnes as he’s entered lots of data).
When you then page to the next page using the arrows the Sales Rep
Filter reverts to ‘All’ again. Consequently you cannot page over for
just Deans Opportunities. The ‘Sort’ also appears to do the same thing
as it reverts to Traffic Light again after paging.
2.       CRM Opportunity. Last Activity. At the moment this is
displaying the First Activity against the Opportunity. It need to
display the Last Activity.
3.       CRM Opportunity. Can you change the Term ‘Next Step’ to
‘Stage’. This will need changing in the Browser, in the Opportunity and
if Possible in the Maintain General Tables (let know if this is a big
problem however).
4.       CRM Opportunity. In the Browser can you swap over the
‘Opportunity Number’ Column with the ‘Last Activity’ Column.
5.       CRM Opportunity. If possible add a Sort of ‘Last Activity’.
6.       CRM Opportunity. Please alter the default filter when you
select the CRM Opportunities to be Descending, not Ascending.
7.       Maintain Customers. Now that we’ve given Sales the Power to add
a Customer we have opened ourselves up to problems! You’ll see they have
entered the same customer 4 times 0020 – 0023. Can you enable the Super
User powers to delete Customers if no transactions (either Issues or
Opportunities) exist against them. In this instance we need to keep 0020
but delete the other 3.
8.